Split out the tree traversal visitor wrapper into its own header.
