+++ src/osvr/Util/CMakeLists.txt
@@ -66,6 +66,7 @@ set(API
     "${HEADER_LOCATION}/TimeValue_fwd.h"
     "${HEADER_LOCATION}/TreeNode.h"
     "${HEADER_LOCATION}/TreeNode_fwd.h"
+    "${HEADER_LOCATION}/TreeTraversalVisitor.h"
     "${HEADER_LOCATION}/UniquePtr.h"
     "${HEADER_LOCATION}/Util.h"
     "${HEADER_LOCATION}/ValueOrRange.h"
